requirements.txt
streamlit==1.28.2
opencv-python-headless==4.11.0.86
pillow==9.5.0
numpy==1.26.4
